fovux-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fovux-mcp"
version = "4.1.0"
description = "Local-first edge-AI computer vision workbench MCP server"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11,<3.14"
authors = [{ name = "Fovux Contributors" }]
keywords = ["computer-vision", "yolo", "mcp", "edge-ai", "ultralytics"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "ultralytics>=8.4.41",
    "onnxruntime>=1.24.4",
    "onnx>=1.21.0",
    "opencv-python-headless>=4.13.0.92",
    "numpy>=2.1,<3",
    "pydantic>=2.13",
    "fastmcp>=3.2,<4",
    "fastapi>=0.136",
    "uvicorn[standard]>=0.45",
    "typer>=0.24",
    "rich>=15.0",
    "sqlalchemy>=2.0.49",
    "pillow>=12.2",
    "imagehash>=4.3.2",
    "platformdirs>=4.9",
    "tomli-w>=1.2",
    "structlog>=25.5",
    "watchfiles>=1.1",
    "httpx>=0.28.1",
    "cryptography>=46.0",
]

[project.optional-dependencies]
gpu = ["onnxruntime-gpu>=1.24.4"]
tensorrt = ["tensorrt>=10.7"]
coreml = ["coremltools>=8.1 ; sys_platform == 'darwin'"]
dev = [
    "pytest>=9.0",
    "pytest-cov>=7.1",
    "pytest-asyncio>=0.24",
    "pytest-xdist>=3.6",
    "pytest-benchmark>=4.0",
    "pytest-timeout>=2.4",
    "ruff>=0.15",
    "mypy>=1.20",
    "bandit>=1.9",
    "pip-audit>=2.10",
    "pre-commit>=4.0",
    "hypothesis>=6.152",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7",
    "mkdocstrings[python]>=1.0",
    "build>=1.2",
    "cyclonedx-bom>=7.3",
    "types-pillow",
]

[project.scripts]
fovux = "fovux.cli:app"

fovux-mcp = "fovux.cli:app"



[project.urls]
Homepage = "https://github.com/oaslananka/fovux"
Documentation = "https://github.com/oaslananka/fovux/tree/main/fovux-mcp/docs"
Repository = "https://github.com/oaslananka/fovux"
Issues = "https://github.com/oaslananka/fovux/issues"
Changelog = "https://github.com/oaslananka/fovux/blob/main/fovux-mcp/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/fovux"]
exclude = ["tests/"]

[tool.hatch.build.targets.sdist]
exclude = [
    "coverage.xml",
    "junit.xml",
    ".coverage",
    ".benchmarks/",
    ".hypothesis/",
    "dist/",
    "site/",
    "htmlcov/",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",    # pycodestyle + pyflakes
    "ANN",             # flake8-annotations
    "D",               # pydocstyle
    "S",               # bandit
    "I",               # isort
    "N",               # pep8-naming
    "UP",              # pyupgrade
    "B",               # flake8-bugbear
    "C4",              # flake8-comprehensions
    "PIE",             # flake8-pie
]
ignore = [
    "D100", "D101", "D102", "D104", "D107",  # missing docstring (relaxed for now)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "ANN", "D"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = false
files = ["src/fovux"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--cov=fovux",
    "--cov-fail-under=90",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--junitxml=junit.xml",
    "--strict-markers",
    "--timeout=60",
    "-x",
]
markers = [
    "network: marks tests that require network access",
    "slow: marks tests that take more than 30 seconds",
    "gpu: marks tests that require a GPU",
    "chaos: marks adversarial/chaos tests",
    "contract: marks API contract tests",
    "benchmark: marks performance benchmarks",
    "security: marks security/pentest related tests",
]

[tool.coverage.run]
source = ["src/fovux"]
omit = [
    "tests/*",
    # Entry points and integration-heavy adapters are validated by focused tests,
    # contract tests, or CI smoke paths; keep the 90% gate on deterministic core code.
    "src/fovux/__main__.py",
    "src/fovux/cli.py",
    "src/fovux/core/dataset_utils.py",
    "src/fovux/core/train_worker.py",
    "src/fovux/core/ultralytics_adapter.py",
    "src/fovux/http/routes.py",
    "src/fovux/tools/benchmark_latency.py",
    "src/fovux/tools/eval_error_analysis.py",
    "src/fovux/tools/export_onnx.py",
    "src/fovux/tools/export_tflite.py",
    "src/fovux/tools/infer_image.py",
    "src/fovux/tools/infer_rtsp.py",
    "src/fovux/tools/sync_to_mlflow.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
